backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
alembic==1.15.2
    # via flask-migrate
amqp==5.3.1
    # via kombu
aniso8601==10.0.0
    # via flask-restful
async-timeout==5.0.1
    # via redis
billiard==4.2.1
    # via celery
blinker==1.9.0
    # via flask
celery==5.5.0
    # via -r requirements.in
certifi==2025.1.31
    # via requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   flask
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
    # via -r requirements.in
flask==3.1.0
    # via
    #   -r requirements.in
    #   flask-cors
    #   flask-jwt-extended
    #   flask-login
    #   flask-migrate
    #   flask-restful
    #   flask-sqlalchemy
    #   flask-wtf
flask-cors==5.0.1
    # via -r requirements.in
flask-jwt-extended==4.7.1
    # via -r requirements.in
flask-login==0.6.3
    # via -r requirements.in
flask-migrate==4.1.0
    # via -r requirements.in
flask-restful==0.3.10
    # via -r requirements.in
flask-sqlalchemy==3.1.1
    # via
    #   -r requirements.in
    #   flask-migrate
flask-wtf==1.2.2
    # via -r requirements.in
greenlet==3.1.1
    # via sqlalchemy
gunicorn==20.1.0
    # added manually here
idna==3.10
    # via
    #   email-validator
    #   requests
itsdangerous==2.2.0
    # via
    #   flask
    #   flask-wtf
jinja2==3.1.6
    # via flask
kombu==5.5.2
    # via celery
mako==1.3.9
    # via alembic
markupsafe==3.0.2
    # via
    #   jinja2
    #   mako
    #   werkzeug
    #   wtforms
numpy==2.2.5
    # via -r requirements.in
prompt-toolkit==3.0.50
    # via click-repl
psycopg2-binary==2.9.3
    #added manually here
pyjwt==2.10.1
    # via flask-jwt-extended
python-dateutil==2.9.0.post0
    # via celery
python-dotenv==1.1.0
    # via -r requirements.in
pytz==2025.2
    # via flask-restful
redis==5.2.1
    # via -r requirements.in
requests==2.32.3
    # via -r requirements.in
six==1.17.0
    # via
    #   flask-restful
    #   python-dateutil
sqlalchemy==2.0.40
    # via
    #   alembic
    #   flask-sqlalchemy
typing-extensions==4.13.0
    # via
    #   alembic
    #   sqlalchemy
tzdata==2025.2
    # via kombu
urllib3==2.3.0
    # via requests
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
werkzeug==3.1.3
    # via
    #   flask
    #   flask-cors
    #   flask-jwt-extended
    #   flask-login
wtforms==3.2.1
    # via flask-wtf
